+++ src/mouseCoordinates.ts
@@ -1,0 +1,5 @@
+import robot from 'robotjs';
+export function getMousePosition() {
+  const currentMousePosition = robot.getMousePos();
+  return { x: currentMousePosition.x, y: currentMousePosition.y };
+}
